isolate top-level-callers
